Move tree hashing code
